Get rid of duplicate progressbar control
