+++ src/Experimental/experimental-gui/Experimental.Gui.csproj
@@ -51,9 +51,6 @@
       <Link>CommonAssemblyInfo.cs</Link>
     </Compile>
     <Compile Include="CommandLineOptions.cs" />
-    <Compile Include="Controls\TestCentricProgressBar.cs">
-      <SubType>Component</SubType>
-    </Compile>
     <Compile Include="Presenters\AddinsPresenter.cs" />
     <Compile Include="Presenters\CategoryGrouping.cs" />
     <Compile Include="Presenters\GroupDisplayStrategy.cs" />
